requirements.txt
# Core dependencies for NeuralTrain Forge
streamlit>=1.28.0
transformers>=4.35.0
peft>=0.6.0
accelerate>=0.24.0
torch>=2.1.0
datasets>=2.14.0
tokenizers>=0.14.0
huggingface-hub>=0.17.0

# Configuration and data handling
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization and monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
tqdm>=4.66.0
psutil>=5.9.0
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0

# Optional GPU support
# torch-audio>=2.1.0  # Uncomment if needed
# torchvision>=0.16.0  # Uncomment if needed

